+++ packages/react-server/examples/basic/src/adapters/node.ts
@@ -1,4 +1,4 @@
 import { createMiddleware } from "@hattip/adapter-node/native-fetch";
-import { handler } from "../entry-ssr";
+import { handler } from "../entry-server";
 
 export default createMiddleware((ctx) => handler(ctx.request));
