+++ packages/react-server/examples/basic/src/entry-server.tsx
@@ -1,0 +1,1 @@
+export { handler } from "@hiogawa/react-server/entry-server";
