--- packages/react-server/examples/basic/src/entry-ssr.tsx
@@ -1,1 +1,0 @@
-export { handler } from "@hiogawa/react-server/entry-ssr";
