--- packages/react-server/src/entry/shared.tsx
@@ -1,6 +1,0 @@
-// TODO: not really shared
-export { createClientReference } from "../lib/rsc";
-export {
-  createServerReference,
-  createServerReferenceForRsc,
-} from "../lib/shared";
